# swagger/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_index","parameters":[],"responses":{"200":{"description":""}}}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginPayload"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/password-reset":{"post":{"operationId":"AuthController_passwordReset","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BasicUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/company/ping":{"get":{"operationId":"CompanyController_index","parameters":[],"responses":{"200":{"description":""}},"tags":["Company"]}},"/api/company":{"get":{"operationId":"CompanyController_getMyCompanies","parameters":[],"responses":{"200":{"description":""}},"tags":["Company"]},"post":{"operationId":"CompanyController_createCompany","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyDto"}}}},"responses":{"201":{"description":""}},"tags":["Company"]}},"/api/company/{id}":{"get":{"operationId":"CompanyController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Company"]},"put":{"operationId":"CompanyController_updateCompany","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCompanyDto"}}}},"responses":{"200":{"description":""}},"tags":["Company"]},"delete":{"operationId":"CompanyController_deleteCompany","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Company"]}},"/api/driver":{"get":{"operationId":"DriverController_getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Driver"}}}}}},"tags":["Driver"]},"post":{"operationId":"DriverController_createDriver","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDriverDto"}}}},"responses":{"201":{"description":""}},"tags":["Driver"]}},"/api/driver/{id}":{"get":{"operationId":"DriverController_getDriverById","parameters":[],"responses":{"200":{"description":""}},"tags":["Driver"]},"put":{"operationId":"DriverController_updateDriver","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDriverDto"}}}},"responses":{"200":{"description":""}},"tags":["Driver"]},"delete":{"operationId":"DriverController_deleteDriver","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Driver"]}},"/api/driver-periodic-inspection":{"get":{"operationId":"DriverPeriodicInspectionController_getInspectionById","parameters":[{"name":"DriverId","required":true,"in":"query","schema":{"type":"number"}},{"name":"FromDate","required":true,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"ToDate","required":true,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"DocumentType","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Driver periodic inspection"]},"post":{"operationId":"DriverPeriodicInspectionController_createInspection","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDriverPeriodicInspectionDto"}}}},"responses":{"201":{"description":""}},"tags":["Driver periodic inspection"]},"put":{"operationId":"DriverPeriodicInspectionController_updateInspection","parameters":[{"name":"DriverId","required":true,"in":"query","schema":{"type":"number"}},{"name":"FromDate","required":true,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"ToDate","required":true,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"DocumentType","required":true,"in":"query","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDriverPeriodicInspectionDto"}}}},"responses":{"200":{"description":""}},"tags":["Driver periodic inspection"]},"delete":{"operationId":"DriverPeriodicInspectionController_deleteInspection","parameters":[{"name":"DriverId","required":true,"in":"query","schema":{"type":"number"}},{"name":"FromDate","required":true,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"ToDate","required":true,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"DocumentType","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Driver periodic inspection"]}},"/api/users/me":{"get":{"operationId":"UsersController_getMe","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/vehicle-mileage/ping":{"get":{"operationId":"VehicleMileageController_ping","parameters":[],"responses":{"200":{"description":""}},"tags":["Vehicle mileage"]}},"/api/vehicle-mileage/newest/{vehicleId}":{"get":{"operationId":"VehicleMileageController_getNewest","parameters":[{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Vehicle mileage"]}},"/api/vehicle-mileage/byVehicle/{vehicleId}":{"get":{"operationId":"VehicleMileageController_getAllByVehicle","parameters":[{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Vehicle mileage"]}},"/api/vehicle-mileage/{vehicleId}/{mileageKm}/{date}":{"get":{"operationId":"VehicleMileageController_getById","parameters":[{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}},{"name":"mileageKm","required":true,"in":"path","schema":{"type":"number"}},{"name":"date","required":true,"in":"path","schema":{"format":"date-time","type":"string"}}],"responses":{"200":{"description":""}},"tags":["Vehicle mileage"]},"delete":{"operationId":"VehicleMileageController_deleteMileage","parameters":[{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}},{"name":"mileageKm","required":true,"in":"path","schema":{"type":"number"}},{"name":"date","required":true,"in":"path","schema":{"format":"date-time","type":"string"}}],"responses":{"200":{"description":""}},"tags":["Vehicle mileage"]}},"/api/vehicle-mileage":{"post":{"operationId":"VehicleMileageController_setMileage","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVehicleMileageDto"}}}},"responses":{"201":{"description":""}},"tags":["Vehicle mileage"]}},"/api/vehicle/ping":{"get":{"operationId":"VehicleController_index","parameters":[],"responses":{"200":{"description":""}},"tags":["Vehicle"]}},"/api/vehicle":{"get":{"operationId":"VehicleController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Vehicle"]},"post":{"operationId":"VehicleController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVehicleDto"}}}},"responses":{"201":{"description":""}},"tags":["Vehicle"]}},"/api/vehicle/{id}":{"get":{"operationId":"VehicleController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Vehicle"]},"put":{"operationId":"VehicleController_updateVehicle","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateVehicleDto"}}}},"responses":{"200":{"description":""}},"tags":["Vehicle"]},"delete":{"operationId":"VehicleController_deleteVehicle","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Vehicle"]}},"/api/vehicle/{id}/exist":{"get":{"operationId":"VehicleController_isExist","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Vehicle"]}},"/api/vehicle-periodic-inspection":{"get":{"operationId":"VehiclePeriodicInspectionController_getAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Vehicle periodic inspection"]},"post":{"operationId":"VehiclePeriodicInspectionController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVehiclePeriodicInspectionDto"}}}},"responses":{"201":{"description":""}},"tags":["Vehicle periodic inspection"]}},"/api/vehicle-periodic-inspection/byVehicle/{vehicleId}":{"get":{"operationId":"VehiclePeriodicInspectionController_getByVehicle","parameters":[{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Vehicle periodic inspection"]}},"/api/vehicle-periodic-inspection/newest/{vehicleId}/{inspectionType}":{"get":{"operationId":"VehiclePeriodicInspectionController_getNewest","parameters":[{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}},{"name":"inspectionType","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Vehicle periodic inspection"]}},"/api/vehicle-periodic-inspection/{vehicleId}/{fromDate}/{toDate}/{inspectionType}":{"get":{"operationId":"VehiclePeriodicInspectionController_getById","parameters":[{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}},{"name":"fromDate","required":true,"in":"path","schema":{}},{"name":"toDate","required":true,"in":"path","schema":{}},{"name":"inspectionType","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Vehicle periodic inspection"]},"put":{"operationId":"VehiclePeriodicInspectionController_update","parameters":[{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}},{"name":"fromDate","required":true,"in":"path","schema":{}},{"name":"toDate","required":true,"in":"path","schema":{}},{"name":"inspectionType","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateVehiclePeriodicInspectionDto"}}}},"responses":{"200":{"description":""}},"tags":["Vehicle periodic inspection"]},"delete":{"operationId":"VehiclePeriodicInspectionController_delete","parameters":[{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}},{"name":"fromDate","required":true,"in":"path","schema":{}},{"name":"toDate","required":true,"in":"path","schema":{}},{"name":"inspectionType","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Vehicle periodic inspection"]}},"/api/driver-preferred-vehicles/all/{byType}/{id}":{"get":{"operationId":"DriverPreferredVehiclesController_getPreferredData","parameters":[{"name":"byType","required":true,"in":"path","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Driver Preferred Vehicles"]}},"/api/driver-preferred-vehicles/{driverId}/{vehicleId}":{"get":{"operationId":"DriverPreferredVehiclesController_getById","parameters":[{"name":"driverId","required":true,"in":"path","schema":{"type":"number"}},{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Driver Preferred Vehicles"]},"put":{"operationId":"DriverPreferredVehiclesController_update","parameters":[{"name":"driverId","required":true,"in":"path","schema":{"type":"number"}},{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDriverPreferredVehiclesDto"}}}},"responses":{"200":{"description":""}},"tags":["Driver Preferred Vehicles"]},"delete":{"operationId":"DriverPreferredVehiclesController_delete","parameters":[{"name":"driverId","required":true,"in":"path","schema":{"type":"number"}},{"name":"vehicleId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Driver Preferred Vehicles"]}},"/api/driver-preferred-vehicles":{"post":{"operationId":"DriverPreferredVehiclesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDriverPreferredVehiclesDto"}}}},"responses":{"201":{"description":""}},"tags":["Driver Preferred Vehicles"]}}},"info":{"title":"Bus Transport","description":"The API of bus transport management app.","version":"0.8","contact":{}},"tags":[{"name":"driver","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginPayload":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"BasicUserDto":{"type":"object","properties":{}},"CreateCompanyDto":{"type":"object","properties":{}},"UpdateCompanyDto":{"type":"object","properties":{}},"Driver":{"type":"object","properties":{"Id":{"type":"number"},"Name":{"type":"string"},"Surname":{"type":"string"},"Phone":{"type":"string","nullable":true},"Email":{"type":"string","nullable":true},"IsPhoneConfirmed":{"type":"boolean","default":false},"IsEmailConfirmed":{"type":"boolean","default":false}},"required":["Id","Name","Surname","Phone","Email","IsPhoneConfirmed","IsEmailConfirmed"]},"CreateDriverDto":{"type":"object","properties":{}},"UpdateDriverDto":{"type":"object","properties":{}},"CreateDriverPeriodicInspectionDto":{"type":"object","properties":{}},"UpdateDriverPeriodicInspectionDto":{"type":"object","properties":{}},"CreateVehicleMileageDto":{"type":"object","properties":{}},"CreateVehicleDto":{"type":"object","properties":{}},"UpdateVehicleDto":{"type":"object","properties":{}},"CreateVehiclePeriodicInspectionDto":{"type":"object","properties":{}},"UpdateVehiclePeriodicInspectionDto":{"type":"object","properties":{}},"CreateDriverPreferredVehiclesDto":{"type":"object","properties":{}},"UpdateDriverPreferredVehiclesDto":{"type":"object","properties":{}}}}}